requirements.txt
aiofiles==24.1.0
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
certifi==2024.6.2
click==8.1.7
colorama==0.4.6
cython==3.0.10
dnspython==2.6.1
ecdsa==0.19.0
email-validator==2.2.0
fastapi==0.110.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
jinja2==3.1.4
mako==1.3.2
markupsafe==2.1.5
mypy==1.10.1
mypy-extensions==1.0.0
orjson==3.10.5
packaging==24.1
passlib==1.7.4
pip==24.0
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.6.4
pydantic-core==2.16.3
pydantic-extra-types==2.8.2
pydantic-settings==2.2.1
pymysql==1.1.1
pytest==8.2.2
pytest-asyncio==0.23.7
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pyyaml==6.0.1
rsa==4.9
ruff==0.3.4
six==1.16.0
sniffio==1.3.1
sqlalchemy==2.0.29
starlette==0.36.3
typing-extensions==4.10.0
ujson==5.10.0
uv==0.2.12
uvicorn==0.29.0
watchfiles==0.22.0
websockets==12.0
